misc_behr_update_plots has method to generate MODIS quality filtered output
This is also the first commit after I split the main BEHR repo up into separate
repositories. This necessitated adding a check for the Git commit of both the
BEHR-core and BEHR-core-utils repositories to the method that actually calls
read_omno2_v_aug2012 and BEHR_main. The public method that generate the data
with quality filtered MODIS albedo has this in its call to that method, but
the old methods that generate the incremental runs 1-5b do not yet. Those
also need their commit hashes updated to represent the new hashed in the post
split up BEHR repo.